+++ src/main/java/fr/ill/ics/cameo/manager/Manager.java
@@ -456,7 +456,8 @@ public class Manager extends ConfigLoader {
 			int state = application.getApplicationState();
 			
 			if (state == ApplicationState.STARTING
-				|| state == ApplicationState.RUNNING) {
+				|| state == ApplicationState.RUNNING
+				|| state == ApplicationState.STOPPING) {
 				return true;
 			}
 			return false;