+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * JZMQ support.
 * Implementation of unmanaged applications.
 * Max applications property is now restricted to the set of application instances with the same name and not the entire set.
+* Implemented the possibility to kill an application that is stopping.
 
 0.0.3
 -----